chore: fix import path after moving UserProvider
