+++ client/facetome/src/components/AllPost.jsx
@@ -6,7 +6,7 @@ import { faCircleNotch } from "@fortawesome/free-solid-svg-icons";
 import DeletePost from "./DeletePost";
 import ToggleLikePost from "./ToggleLikePost";
 import DeleteComment from "./DeleteComment";
-import { useUser } from "../context/UserContext";
+import { useUser } from "../provider/UserProvider";
 
 const AllPost = ({
   posts,
