+++ client/facetome/src/main.jsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import "./styles/index.css";
 import AppRoutes from "./AppRoutes.jsx";
-import { UserProvider } from "./context/UserContext.jsx";
+import { UserProvider } from "./provider/UserProvider.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
